third_party/blink/web_tests/fast/dom/shadow/shadow-hierarchy-exception-expected.txt

Appending an ancestor to a shadow tree should throw an exception

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS shadowDiv.appendChild(container) threw exception HierarchyRequestError: Failed to execute 'appendChild' on 'Node': The new child element contains the parent..
PASS successfullyParsed is true

TEST COMPLETE
